+++ kafka-webview-ui/src/main/java/org/sourcelab/kafka/webview/ui/manager/ui/datatable/DatatableColumn.java
@@ -1,11 +1,13 @@
-package org.sourcelab.kafka.webview.ui.manager.datatable;
+package org.sourcelab.kafka.webview.ui.manager.ui.datatable;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Function;
 
 /**
- *
+ * Defines a column within a datatable.
+ * @param <T> Type of object being rendered by the column.
  */
 public class DatatableColumn<T> {
     private final String fieldName;
@@ -15,19 +17,39 @@ public class DatatableColumn<T> {
     private Function<T, String> renderFunction;
     private final RenderTemplate renderTemplate;
 
+    /**
+     * Constructor.
+     * See Builder instance.
+     *
+     * @param fieldName The underlying field name to be displayed in the column.
+     * @param label Display label for the column.
+     * @param colSpan How many columns should the field span.
+     * @param isSortable Is the column sortable.
+     * @param renderFunction A callable function that takes in the current record and returns the value to be displayed.
+     * @param renderTemplate A callabale thymeleaf template.
+     */
     public DatatableColumn(
         final String fieldName, final String label, final int colSpan, final boolean isSortable,
         final Function<T, String> renderFunction,
         final RenderTemplate renderTemplate
     ) {
-        this.fieldName = fieldName;
-        this.label = label;
+        this.fieldName = Objects.requireNonNull(fieldName);
+        this.label = Objects.requireNonNull(label);
         this.colSpan = colSpan;
         this.isSortable = isSortable;
+
+        // One of these may not be null.
         this.renderFunction = renderFunction;
         this.renderTemplate = renderTemplate;
     }
 
+    /**
+     * The RenderTemplate instance used to render the columns values.
+     * If no specific instance was provided to the constructor, the default render template
+     * will be created and used.
+     *
+     * @return RenderTemplate.
+     */
     public RenderTemplate getRenderTemplate() {
         if (renderTemplate == null) {
             return new RenderTemplate<T>("fragments/datatable/fields/TextValue", "display") {
@@ -40,6 +62,11 @@ public class DatatableColumn<T> {
         return renderTemplate;
     }
 
+    /**
+     * Render the current columns value for the given record.
+     * @param record The record to render the column field value for.
+     * @return Rendered field value.
+     */
     public String render(final T record) {
         return renderFunction.apply(record);
     }
@@ -60,6 +87,12 @@ public class DatatableColumn<T> {
         return isSortable;
     }
 
+    /**
+     * Create a new Builder instance.
+     * @param type The type of record being rendered.
+     * @param <T> The type of record being rendered.
+     * @return New Builder instance.
+     */
     public static <T> Builder<T> newBuilder(Class<T> type) {
         return new Builder<T>();
     }
@@ -72,7 +105,10 @@ public class DatatableColumn<T> {
             + '}';
     }
 
-
+    /**
+     * Builder instance for DatatableColumn.
+     * @param <T> Type of object being rendered in the column.
+     */
     public static final class Builder<T> {
         private String fieldName;
         private String label;
@@ -114,6 +150,10 @@ public class DatatableColumn<T> {
             return this;
         }
 
+        /**
+         * Create new DatatableColumn from builder.
+         * @return new DatatableColumn.
+         */
         public DatatableColumn<T> build() {
             return new DatatableColumn<T>(fieldName, label, colSpan, isSortable, renderFunction, renderTemplate);
         }
