+++ kafka-webview-ui/src/main/java/org/sourcelab/kafka/webview/ui/manager/ui/datatable/DatatableSearch.java
@@ -1,4 +1,4 @@
-package org.sourcelab.kafka.webview.ui.manager.datatable;
+package org.sourcelab.kafka.webview.ui.manager.ui.datatable;
 
 /**
  *
@@ -8,6 +8,12 @@ public class DatatableSearch {
     private final String field;
     private final String currentSearchTerm;
 
+    /**
+     * Constructor.
+     * @param label Human readable display label.
+     * @param field The underlying field to search over.
+     * @param currentSearchTerm The current search term if defined in a request.
+     */
     public DatatableSearch(final String label, final String field, final String currentSearchTerm) {
         this.label = label;
         this.field = field;
