--- kafka-webview-ui/src/main/java/org/sourcelab/kafka/webview/ui/manager/ui/datatable/Direction.java
@@ -1,6 +1,0 @@
-package org.sourcelab.kafka.webview.ui.manager.ui.datatable;
-
-public enum Direction {
-    asc,
-    desc;
-}
